refactor(sessions): isolate queued transcript context
